Fix bug when rotating near edge or landed blocks
There was no bounds checking or collision check when rotating so it was
possible to rotate and end up partially offscreen or rotate into already
landed blocks.

Required a change of the Grid model to make it cleaner but I am also
able to reuse the same bounds checking for moving and rotating.